requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
ultralytics>=8.4.22
pytesseract>=0.3.10
pyyaml>=6.0
scikit-learn>=1.3.0
sklearn>=1.3.0